moved UI based launcher to org.palladiosimulator.simexp.ui.workflow
