+++ bundles/org.palladiosimulator.simexp.ui.workflow/plugin.xml
@@ -4,7 +4,7 @@
    <extension
          point="org.eclipse.debug.core.launchConfigurationTypes">
       <launchConfigurationType
-            delegate="org.palladiosimulator.simexp.workflow.launcher.SimExpLauncher"
+            delegate="org.palladiosimulator.simexp.ui.workflow.launcher.SimExpUILauncher"
             id="org.palladiosimulator.simexp.ui.workflow.launchConfigurationType1"
             modes="run, debug"
             name="SimExp Analyzer"
