+++ bundles/org.palladiosimulator.simexp.ui.workflow/src/main/java/org/palladiosimulator/simexp/ui/workflow/launcher/SimExpUILauncher.java
@@ -1,0 +1,29 @@
+package org.palladiosimulator.simexp.ui.workflow.launcher;
+
+import org.eclipse.core.runtime.CoreException;
+import org.eclipse.core.runtime.IProgressMonitor;
+import org.eclipse.debug.core.ILaunch;
+import org.palladiosimulator.simexp.workflow.api.SimExpWorkflowConfiguration;
+import org.palladiosimulator.simexp.workflow.launcher.SimExpLauncher;
+
+import de.uka.ipd.sdq.workflow.BlackboardBasedWorkflow;
+import de.uka.ipd.sdq.workflow.WorkflowExceptionHandler;
+import de.uka.ipd.sdq.workflow.mdsd.blackboard.MDSDBlackboard;
+import de.uka.ipd.sdq.workflow.ui.UIBasedWorkflow;
+import de.uka.ipd.sdq.workflow.ui.UIBasedWorkflowExceptionHandler;
+
+public class SimExpUILauncher extends SimExpLauncher {
+    @Override
+    protected BlackboardBasedWorkflow<MDSDBlackboard> createWorkflow(
+            final SimExpWorkflowConfiguration workflowConfiguration, final IProgressMonitor monitor,
+            final ILaunch launch) throws CoreException {
+        return new UIBasedWorkflow<>(this.createWorkflowJob(workflowConfiguration, launch), monitor,
+                this.createExceptionHandler(workflowConfiguration.isInteractive()), this.createBlackboard());
+    }
+
+    @Override
+    protected WorkflowExceptionHandler createExceptionHandler(boolean interactive) {
+        return new UIBasedWorkflowExceptionHandler(!interactive);
+    }
+
+}
